+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Allow `assertionMethod` property value to be object.
+- Attempt to load additional page when importing replies from Mastodon.
 
 ## [2.4.1] - 2023-11-23